chore: refactor EventProcessorRuntime into its own crate
